projects: eval-adis: Rename eval-adis to eval-adis1650x
Rename eval-adis to eval-adis1650x for clarity.
This project supports examples with adis1650x devices.
