+++ projects/eval-adis1650x/src.mk
