+++ projects/eval-adis1650x/src/examples/examples_src.mk
