+++ projects/eval-adis1650x/src/examples/iio_trigger_example/iio_trigger_example.c
@@ -1,6 +1,6 @@
 /***************************************************************************//**
  *   @file   iio_trigger_example.c
- *   @brief  Implementation of IIO trigger example for eval-adis project.
+ *   @brief  Implementation of IIO trigger example for eval-adis1650x project.
  *   @author RBolboac (ramona.bolboaca@analog.com)
 ********************************************************************************
  * Copyright 2022(c) Analog Devices, Inc.
