+++ projects/eval-adis1650x/src/platform/maxim/main.c
@@ -1,6 +1,6 @@
 /***************************************************************************//**
- *   @file   dummy_example.h
- *   @brief  DUMMY example header for eval-adis project
+ *   @file   main.c
+ *   @brief  Main file for Maxim platform of eval-adis1650x project.
  *   @author RBolboac (ramona.bolboaca@analog.com)
 ********************************************************************************
  * Copyright 2023(c) Analog Devices, Inc.
@@ -37,13 +37,92 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
 
-#ifndef __DUMMY_EXAMPLE_H__
-#define __DUMMY_EXAMPLE_H__
-
 /******************************************************************************/
-/************************ Functions Declarations ******************************/
+/***************************** Include Files **********************************/
 /******************************************************************************/
 
-int dummy_example_main();
+#include "platform_includes.h"
+#include "common_data.h"
+#include "no_os_error.h"
 
-#endif /* __DUMMY_EXAMPLE_H__ */
+#ifdef IIO_TRIGGER_EXAMPLE
+#include "iio_trigger_example.h"
+#endif
+
+#ifdef DUMMY_EXAMPLE
+#include "dummy_example.h"
+#endif
+
+/******************************************************************************/
+/************************* Functions Definitions ******************************/
+/******************************************************************************/
+
+/**
+ * @brief Main function execution for STM32 platform.
+ *
+ * @return ret - Result of the enabled examples execution.
+ */
+int main()
+{
+	int ret = -EINVAL;
+
+	adis1650x_ip.spi_init = &adis1650x_spi_ip;
+
+#ifdef DUMMY_EXAMPLE
+	struct no_os_uart_desc *uart_desc;
+
+	ret = no_os_uart_init(&uart_desc, &adis1650x_uart_ip);
+	if (ret)
+		return ret;
+
+	no_os_uart_stdio(uart_desc);
+	ret = dummy_example_main();
+	if (ret)
+		no_os_uart_remove(uart_desc);
+#endif
+
+#ifdef IIO_TRIGGER_EXAMPLE
+	struct no_os_gpio_desc *adis_gpio_desc;
+	struct no_os_irq_ctrl_desc *nvic_desc;
+	struct no_os_irq_init_param nvic_ip = {
+		.platform_ops = &max_irq_ops,
+	};
+
+	/* Initialize DATA READY pin */
+	ret = no_os_gpio_get_optional(&adis_gpio_desc, &adis_gpio_drdy_ip);
+	if (ret)
+		return ret;
+
+	ret = no_os_gpio_direction_input(adis_gpio_desc);
+	if (ret)
+		goto error_gpio;
+
+	/* Initialize GPIO IRQ controller */
+	ret = no_os_irq_ctrl_init(&nvic_desc, &nvic_ip);
+	if (ret)
+		goto error_gpio;
+
+	ret = no_os_irq_enable(nvic_desc, NVIC_GPIO_IRQ);
+	if (ret)
+		goto error_irq;
+
+	ret = iio_trigger_example_main();
+	if (ret)
+		goto error_irq;
+
+error_irq:
+	no_os_irq_ctrl_remove(nvic_desc);
+error_gpio:
+	no_os_gpio_remove(adis_gpio_desc);
+#endif
+
+#if (DUMMY_EXAMPLE + IIO_TRIGGER_EXAMPLE == 0)
+#error At least one example has to be selected using y value in Makefile.
+#elif (DUMMY_EXAMPLE + IIO_TRIGGER_EXAMPLE > 1)
+#error Selected example projects cannot be enabled at the same time. \
+Please enable only one example and re-build the project.
+#endif
+
+	return ret;
+}
+
