+++ projects/eval-adis1650x/src/platform/maxim/parameters.c
@@ -1,6 +1,6 @@
 /***************************************************************************//**
- *   @file   main.c
- *   @brief  Main file for STM32 platform of eval-adis project.
+ *   @file   parameters.c
+ *   @brief  Definition of Maxim platform data used by eval-adis1650x project.
  *   @author RBolboac (ramona.bolboaca@analog.com)
 ********************************************************************************
  * Copyright 2023(c) Analog Devices, Inc.
@@ -41,58 +41,37 @@
 /***************************** Include Files **********************************/
 /******************************************************************************/
 
-#include "platform_includes.h"
-#include "common_data.h"
-#include "no_os_error.h"
-
-#ifdef IIO_TRIGGER_EXAMPLE
-#include "iio_trigger_example.h"
-#endif
-
-#ifdef DUMMY_EXAMPLE
-#include "dummy_example.h"
-#endif
+#include "parameters.h"
 
 /******************************************************************************/
-/************************* Functions Definitions ******************************/
+/********************** Macros and Constants Definitions **********************/
 /******************************************************************************/
 
-/**
- * @brief Main function execution for STM32 platform.
- *
- * @return ret - Result of the enabled examples execution.
- */
-int main()
-{
-	int ret = -EINVAL;
-	adis1650x_spi_extra_ip.get_input_clock = HAL_RCC_GetPCLK1Freq;
-	adis1650x_ip.spi_init = &adis1650x_spi_ip;
+struct max_uart_init_param adis1650x_uart_extra_ip = {
+	.flow = UART_FLOW_DIS
+};
 
-	stm32_init();
+struct max_gpio_init_param adis1650x_gpio_extra_ip = {
+	.vssel = MXC_GPIO_VSSEL_VDDIOH,
+};
+
+struct max_spi_init_param adis1650x_spi_extra_ip  = {
+	.num_slaves = 1,
+	.polarity = SPI_SS_POL_LOW,
+	.vssel = MXC_GPIO_VSSEL_VDDIOH,
+};
 
 #ifdef IIO_TRIGGER_EXAMPLE
-	ret = iio_trigger_example_main();
+/* Initialization for Sync pin GPIO. */
+struct no_os_gpio_init_param adis_gpio_drdy_ip = {
+	.port = GPIO_DRDY_PORT_NUM,
+	.number = GPIO_DRDY_PIN_NUM,
+	.pull = NO_OS_PULL_NONE,
+	.platform_ops = GPIO_OPS,
+	.extra = GPIO_EXTRA
+};
+
+struct max_gpio_init_param adis_gpio_drdy_extra_ip = {
+	.vssel = MXC_GPIO_VSSEL_VDDIOH,
+};
 #endif
-
-#ifdef DUMMY_EXAMPLE
-	struct no_os_uart_desc *uart_desc;
-
-	ret = no_os_uart_init(&uart_desc, &adis1650x_uart_ip);
-	if (ret)
-		return ret;
-
-	no_os_uart_stdio(uart_desc);
-	ret = dummy_example_main();
-	if (ret)
-		no_os_uart_remove(uart_desc);
-#endif
-
-#if (DUMMY_EXAMPLE + IIO_TRIGGER_EXAMPLE == 0)
-#error At least one example has to be selected using y value in Makefile.
-#elif (DUMMY_EXAMPLE + IIO_TRIGGER_EXAMPLE > 1)
-#error Selected example projects cannot be enabled at the same time. \
-Please enable only one example and re-build the project.
-#endif
-
-	return ret;
-}
