+++ projects/eval-adis1650x/src/platform/maxim/parameters.h
@@ -1,6 +1,7 @@
 /***************************************************************************//**
- *   @file   eval-adis/src/common/common_data.h
- *   @brief  Defines common data to be used by eval-adis examples.
+ *   @file   parameters.h
+ *   @brief  Definitions specific to Maxim platform used by eval-adis1650x
+ *           project.
  *   @author RBolboac (ramona.bolboaca@analog.com)
 ********************************************************************************
  * Copyright 2023(c) Analog Devices, Inc.
@@ -37,37 +38,67 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
 
-#ifndef __COMMON_DATA_H__
-#define __COMMON_DATA_H__
+#ifndef __PARAMETERS_H__
+#define __PARAMETERS_H__
 
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
 
-#include "platform_includes.h"
-#include "adis1650x.h"
-#ifdef IIO_SUPPORT
-#include "iio_adis1650x.h"
-#include "iio_adis.h"
-#ifdef IIO_TRIGGER_EXAMPLE
-#include "iio_trigger.h"
-#endif
-#endif
+#include "maxim_irq.h"
+#include "maxim_spi.h"
+#include "maxim_gpio.h"
+#include "maxim_uart.h"
+#include "maxim_uart_stdio.h"
 
 /******************************************************************************/
 /********************** Macros and Constants Definitions **********************/
 /******************************************************************************/
 
-extern struct no_os_uart_init_param adis1650x_uart_ip;
-extern struct no_os_spi_init_param adis1650x_spi_ip;
-extern struct no_os_gpio_init_param adis1650x_gpio_reset_ip;
-extern struct adis_init_param adis1650x_ip;
-
-#ifdef IIO_TRIGGER_EXAMPLE
-#define ADIS1650X_GPIO_TRIG_NAME "adis16505-2-dev0"
-
-extern struct iio_hw_trig_init_param adis1650x_gpio_trig_ip;
-extern struct no_os_irq_init_param adis1650x_gpio_irq_ip;
+#ifdef IIO_SUPPORT
+#define INTC_DEVICE_ID  0
 #endif
 
-#endif /* __COMMON_DATA_H__ */
+#define UART_DEVICE_ID  0
+#define UART_BAUDRATE   57600
+#define UART_EXTRA      &adis1650x_uart_extra_ip
+#define UART_OPS        &max_uart_ops
+#define UART_IRQ_ID     UART0_IRQn
+
+
+#define SPI_DEVICE_ID   1
+#define SPI_CS          1
+
+#define SPI_BAUDRATE    1000000
+#define SPI_OPS         &max_spi_ops
+#define SPI_EXTRA       &adis1650x_spi_extra_ip
+
+#define GPIO_OPS            &max_gpio_ops
+#define GPIO_EXTRA          &adis1650x_gpio_extra_ip
+
+extern struct max_uart_init_param adis1650x_uart_extra_ip;
+extern struct max_spi_init_param adis1650x_spi_extra_ip;
+extern struct max_gpio_init_param adis1650x_gpio_extra_ip;
+
+#define GPIO_RESET_PIN_NUM   19
+#define GPIO_RESET_PORT_NUM  0
+
+
+#ifdef IIO_TRIGGER_EXAMPLE
+#define GPIO_DRDY_PIN_NUM   6
+#define GPIO_DRDY_PORT_NUM  1
+
+#define NVIC_GPIO_IRQ   GPIO1_IRQn
+
+#define ADIS1650X_GPIO_TRIG_IRQ_ID    GPIO_DRDY_PIN_NUM
+#define ADIS1650X_GPIO_CB_HANDLE      MXC_GPIO_GET_GPIO(GPIO_DRDY_PORT_NUM)
+
+#define GPIO_IRQ_ID             GPIO_DRDY_PORT_NUM
+#define GPIO_IRQ_OPS            &max_gpio_irq_ops
+#define GPIO_IRQ_EXTRA          &adis_gpio_drdy_extra_ip
+
+extern struct no_os_gpio_init_param adis_gpio_drdy_ip;
+extern struct max_gpio_init_param adis_gpio_drdy_extra_ip;
+#endif
+
+#endif /* __PARAMETERS_H__ */
