+++ projects/eval-adis1650x/src/platform/pico/main.c
@@ -1,6 +1,6 @@
 /***************************************************************************//**
- *   @file   parameters.c
- *   @brief  Definition of pico platform data used by eval-adis project.
+ *   @file   main.c
+ *   @brief  Main file for pico platform of eval-adis1650x project.
  *   @author RBolboac (ramona.bolboaca@analog.com)
 ********************************************************************************
  * Copyright 2023(c) Analog Devices, Inc.
@@ -41,20 +41,55 @@
 /***************************** Include Files **********************************/
 /******************************************************************************/
 
-#include "parameters.h"
+#include "platform_includes.h"
+#include "common_data.h"
+#include "no_os_error.h"
+
+#ifdef IIO_TRIGGER_EXAMPLE
+#include "iio_trigger_example.h"
+#endif
+
+#ifdef DUMMY_EXAMPLE
+#include "dummy_example.h"
+#endif
 
 /******************************************************************************/
-/********************** Macros and Constants Definitions **********************/
+/************************* Functions Definitions ******************************/
 /******************************************************************************/
 
-struct pico_spi_init_param adis1650x_spi_extra_ip  = {
-	.spi_tx_pin = SPI0_TX_GP19,
-	.spi_rx_pin = SPI0_RX_GP16,
-	.spi_sck_pin = SPI0_SCK_GP18,
-	.spi_cs_pin = SPI0_CS_GP17
-};
+/**
+ * @brief Main function execution for pico platform.
+ *
+ * @return ret - Result of the enabled examples execution.
+ */
+int main()
+{
+	int ret = -EINVAL;
+	adis1650x_ip.spi_init = &adis1650x_spi_ip;
 
-struct pico_uart_init_param adis1650x_uart_extra_ip = {
-	.uart_tx_pin = UART_TX_PIN,
-	.uart_rx_pin = UART_RX_PIN,
-};
+#ifdef DUMMY_EXAMPLE
+	struct no_os_uart_desc *uart_desc;
+
+	ret = no_os_uart_init(&uart_desc, &adis1650x_uart_ip);
+	if (ret)
+		return ret;
+
+	ret = dummy_example_main();
+	if (ret)
+		no_os_uart_remove(uart_desc);
+#endif
+
+#ifdef IIO_TRIGGER_EXAMPLE
+	ret = iio_trigger_example_main();
+#endif
+
+
+#if (DUMMY_EXAMPLE + IIO_TRIGGER_EXAMPLE == 0)
+#error At least one example has to be selected using y value in Makefile.
+#elif (DUMMY_EXAMPLE + IIO_TRIGGER_EXAMPLE > 1)
+#error Selected example projects cannot be enabled at the same time. \
+Please enable only one example and re-build the project.
+#endif
+
+	return ret;
+}
