+++ projects/eval-adis1650x/src/platform/stm32/parameters.c
@@ -1,9 +1,9 @@
 /***************************************************************************//**
- *   @file   iio_trigger_example.h
- *   @brief  IIO trigger example header for eval-adis project.
+ *   @file   parameters.c
+ *   @brief  Definition of STM32 platform data used by eval-adis1650x project.
  *   @author RBolboac (ramona.bolboaca@analog.com)
 ********************************************************************************
- * Copyright 2022(c) Analog Devices, Inc.
+ * Copyright 2023(c) Analog Devices, Inc.
  *
  * All rights reserved.
  *
@@ -36,16 +36,33 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
-#ifndef __IIO_TRIGGER_EXAMPLE_H__
-#define __IIO_TRIGGER_EXAMPLE_H__
 
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
 
-/******************************************************************************/
-/************************ Functions Declarations ******************************/
-/******************************************************************************/
-int iio_trigger_example_main();
+#include "parameters.h"
 
-#endif /* __IIO_TRIGGER_EXAMPLE_H__ */
+/******************************************************************************/
+/********************** Macros and Constants Definitions **********************/
+/******************************************************************************/
+
+struct stm32_uart_init_param adis1650x_uart_extra_ip = {
+	.huart = &huart5,
+};
+
+
+struct stm32_spi_init_param adis1650x_spi_extra_ip  = {
+	.chip_select_port = SPI_CS_PORT,
+};
+
+struct stm32_gpio_init_param adis1650x_gpio_reset_extra_ip = {
+	.mode = GPIO_MODE_OUTPUT_OD,
+	.speed = GPIO_SPEED_FREQ_VERY_HIGH,
+};
+
+#ifdef IIO_TRIGGER_EXAMPLE
+struct stm32_gpio_irq_init_param adis1650x_gpio_irq_extra_ip = {
+	.port_nb = 0, /* Port A */
+};
+#endif
